added demo atomic structure on the project
